split packets
